requirements.txt
appdirs==1.4.4
asgiref==3.3.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-ckeditor==6.0.0
django-crispy-forms==1.11.1
django-js-asset==1.2.2
django-phonenumber-field==5.0.0
django-utils-six==2.0
filelock==3.0.12
gunicorn==20.1.0
phonenumbers==8.12.19
Pillow==8.1.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0